requirements.txt
streamlit>=1.36
tensorflow-cpu>=2.15
numpy>=1.24
pillow>=10.0
gdown>=5.1
